webuniversity, commit 4

+++ src/main/java/ua/com/foxminded/webuniversity/service/GroupServiceImpl.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import ua.com.foxminded.webuniversity.dao.EntityNotFoundException;
 import ua.com.foxminded.webuniversity.dao.GroupRepository;
 import ua.com.foxminded.webuniversity.entity.Group;
+import ua.com.foxminded.webuniversity.exception.EntityNotFoundException;
 
 @Service
 public class GroupServiceImpl implements GroupService {
@@ -42,7 +42,12 @@ public class GroupServiceImpl implements GroupService {
     }
 
     public void delete(Integer id) {
-        groupRepository.deleteById(id);
+        Optional<Group> result = groupRepository.findById(id);
+        if (result.isPresent()) {
+            groupRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Cannot find group id = " + id);
+        }
     }
     
 }
